backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api>=3.1.11",
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "svix>=1.73.0",
    "uvicorn>=0.35.0",
]
